web/requirements.txt
diff-match-patch==20121119
Django==1.9.2
django-allauth==0.24.1
django-bootstrap3==6.2.2
django-crispy-forms==1.6.0
django-import-export==0.4.2
django-paintstore==0.2
django-suit==0.2.16
django-widget-tweaks==1.4.1
djrill==2.0
factory-boy==2.6.1
fake-factory==0.5.3
gunicorn==19.4.5
oauthlib==1.0.3
psycopg2==2.6.1
python-openid==2.2.5
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.6.0
tablib==0.11.1
virtualenv==14.0.6
wheel==0.29.0
